organized all of the buttons, events, utilities, and html structure into folders and created import and exports to clean up my index.html
