+++ utils/data/houses.js
@@ -1,0 +1,24 @@
+const houses = [
+  {
+    house: 'gryffindor',
+    crest:
+      'https://static.wikia.nocookie.net/pottermore/images/1/16/Gryffindor_crest.png'
+  },
+  {
+    house: 'slytherin',
+    crest:
+      'https://static.wikia.nocookie.net/pottermore/images/4/45/Slytherin_Crest.png'
+  },
+  {
+    house: 'hufflepuff',
+    crest:
+      'https://static.wikia.nocookie.net/pottermore/images/5/5e/Hufflepuff_crest.png'
+  },
+  {
+    house: 'ravenclaw',
+    crest:
+      'https://static.wikia.nocookie.net/pottermore/images/4/4f/Ravenclaw_crest.png'
+  }
+];
+
+export default houses;
